requirements.txt
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
discord==1.0.1
discord.py==1.6.0
et-xmlfile==1.0.1
future==0.18.2
Glances==3.2.2
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
openpyxl==3.0.5
psutil==5.8.0
pylint==2.6.0
PyPDF2==1.26.0
python-dotenv==0.14.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.5
wrapt==1.12.1
xlrd==1.2.0
yarl==1.6.3
aiohttp~=3.8.1
asyncer~=0.0.2
click~=8.1.3
fastapi~=0.87.0
filetype~=1.2.0
pooch~=1.6.0
imagehash~=4.3.1
numpy~=1.23.5
onnxruntime~=1.13.1
opencv-python-headless~=4.6.0.66
pillow~=9.3.0
pymatting~=1.1.8
python-multipart~=0.0.5
scikit-image~=0.19.3
scipy~=1.9.3
tqdm~=4.64.1
uvicorn~=0.20.0
watchdog~=2.1.9
